requirements.txt
yfinance
duckdb
pandas
pyarrow
streamlit
apache-airflow
pyyaml
pytest